+++ examples/custom_frontend_options.cpp
@@ -14,7 +14,7 @@
 struct CustomFrontendOptions
 {
   static constexpr quill::QueueType queue_type = quill::QueueType::BoundedDropping;
-  static constexpr uint32_t initial_queue_capacity = 131'072;
+  static constexpr size_t initial_queue_capacity = 131'072;
   static constexpr uint32_t blocking_queue_retry_interval_ns = 800;
   static constexpr quill::HugePagesPolicy huge_pages_policy = quill::HugePagesPolicy::Never;
 };