+++ include/quill/Logger.h
@@ -51,7 +51,6 @@ public:
   using frontend_options_t = TFrontendOptions;
 
   static constexpr bool using_unbounded_queue =
-    (frontend_options_t::queue_type == QueueType::UnboundedUnlimited) ||
     (frontend_options_t::queue_type == QueueType::UnboundedBlocking) ||
     (frontend_options_t::queue_type == QueueType::UnboundedDropping);
 
@@ -140,13 +139,8 @@ public:
 
     std::byte* write_buffer = _prepare_write_buffer(total_size);
 
-    if constexpr (frontend_options_t::queue_type == QueueType::UnboundedUnlimited)
-    {
-      assert(write_buffer &&
-             "Unbounded unlimited queue will always allocate and have enough capacity");
-    }
-    else if constexpr ((frontend_options_t::queue_type == QueueType::BoundedDropping) ||
-                       (frontend_options_t::queue_type == QueueType::UnboundedDropping))
+    if constexpr ((frontend_options_t::queue_type == QueueType::BoundedDropping) ||
+                  (frontend_options_t::queue_type == QueueType::UnboundedDropping))
     {
       if (QUILL_UNLIKELY(write_buffer == nullptr))
       {
@@ -370,10 +364,10 @@ private:
       // MSVC doesn't like the template keyword, but every other compiler requires it
 #if defined(_MSC_VER)
       return thread_context->get_spsc_queue<frontend_options_t::queue_type>().prepare_write(
-        total_size, frontend_options_t::queue_type);
+        total_size, frontend_options_t::unbounded_queue_max_capacity);
 #else
       return thread_context->get_spsc_queue<frontend_options_t::queue_type>()
-        .template prepare_write<frontend_options_t::queue_type>(total_size);
+        .template prepare_write<frontend_options_t::unbounded_queue_max_capacity>(total_size);
 #endif
     }
     else