+++ include/quill/core/UnboundedSPSCQueue.h
@@ -107,9 +107,9 @@ public:
    */
 #if defined(_MSC_VER)
   // MSVC doesn't like this as template <QueueType queue_type> when called from Logger, while it compiles on MSVC there will be false positives from clang-tidy
-  QUILL_NODISCARD QUILL_ATTRIBUTE_HOT std::byte* prepare_write(size_t nbytes, QueueType queue_type)
+  QUILL_NODISCARD QUILL_ATTRIBUTE_HOT std::byte* prepare_write(size_t nbytes, size_t unbounded_queue_max_capacity)
 #else
-  template <QueueType queue_type>
+  template <size_t unbounded_queue_max_capacity>
   QUILL_NODISCARD QUILL_ATTRIBUTE_HOT std::byte* prepare_write(size_t nbytes)
 #endif
   {
@@ -122,9 +122,9 @@ public:
     }
 
 #if defined(_MSC_VER)
-    return _handle_full_queue(nbytes, queue_type);
+    return _handle_full_queue(nbytes, unbounded_queue_max_capacity);
 #else
-    return _handle_full_queue<queue_type>(nbytes);
+    return _handle_full_queue<unbounded_queue_max_capacity>(nbytes);
 #endif
   }
 
@@ -242,57 +242,45 @@ public:
     return _consumer->bounded_queue.empty() && (_consumer->next.load(std::memory_order_relaxed) == nullptr);
   }
 
-
 private:
   /***/
 #if defined(_MSC_VER)
-  QUILL_NODISCARD std::byte* _handle_full_queue(size_t nbytes, QueueType queue_type)
+  QUILL_NODISCARD std::byte* _handle_full_queue(size_t nbytes, size_t unbounded_queue_max_capacity)
 #else
-  template <QueueType queue_type>
+  template <size_t unbounded_queue_max_capacity>
   QUILL_NODISCARD std::byte* _handle_full_queue(size_t nbytes)
 #endif
   {
     // Then it means the queue doesn't have enough size
     size_t capacity = _producer->bounded_queue.capacity() * 2ull;
-    while (capacity < (nbytes + 1))
+    while (capacity < nbytes)
     {
       capacity = capacity * 2ull;
     }
 
-#if defined(_MSC_VER)
-    if ((queue_type == QueueType::UnboundedBlocking) || (queue_type == QueueType::UnboundedDropping))
-#else
-    if constexpr ((queue_type == QueueType::UnboundedBlocking) || (queue_type == QueueType::UnboundedDropping))
-#endif
+    if (QUILL_UNLIKELY(capacity > unbounded_queue_max_capacity))
     {
-      size_t constexpr max_bounded_queue_size = 2ull * 1024 * 1024 * 1024; // 2 GB
-
-      if (QUILL_UNLIKELY(capacity > max_bounded_queue_size))
+      if (nbytes > unbounded_queue_max_capacity)
       {
-        if (nbytes > max_bounded_queue_size)
-        {
-          QUILL_THROW(QuillError{
-            "Logging single messages larger than 2 GB is not supported with the current queue "
-            "type. For UnboundedBlocking or UnboundedDropping queues, this limitation applies.\n"
-            "To log single messages larger than 2 GB, consider using the UnboundedUnlimited queue "
-            "type.\n"
-            "Message size: " +
-            std::to_string(nbytes) +
-            " bytes\n"
-            "Required queue capacity: " +
-            std::to_string(capacity) +
-            " bytes\n"
-            "Maximum allowed queue capacity: " +
-            std::to_string(max_bounded_queue_size) + " bytes"});
-        }
-
-        // we reached the max_bounded_queue_size we won't be allocating more
-        // instead return nullptr to block or drop
-        return nullptr;
+        QUILL_THROW(
+          QuillError{"Logging single messages larger than the configured maximum queue capacity "
+                     "is not possible.\n"
+                     "To log single messages exceeding this limit, consider increasing "
+                     "FrontendOptions::unbounded_queue_max_capacity.\n"
+                     "Message size: " +
+                     std::to_string(nbytes) +
+                     " bytes\n"
+                     "Required queue capacity: " +
+                     std::to_string(capacity) +
+                     " bytes\n"
+                     "Configured maximum queue capacity: " +
+                     std::to_string(unbounded_queue_max_capacity) + " bytes"});
       }
-    }
 
-    // else the UnboundedUnlimited queue has no limits
+      // we reached the unbounded_queue_max_capacity we won't be allocating more
+      // instead return nullptr to block or drop
+      return nullptr;
+    }
 
     // commit previous write to the old queue before switching
     _producer->bounded_queue.commit_write();