+++ include/quill/sinks/SyslogSink.h
@@ -154,12 +154,10 @@ public:
   /**
    * @brief Writes a formatted log message to the stream
    */
-  void write_log(MacroMetadata const* /* log_metadata */,
-                 uint64_t /* log_timestamp */, std::string_view /* thread_id */,
-                 std::string_view /* thread_name */, std::string const& /* process_id */,
-                 std::string_view /* logger_name */, LogLevel log_level,
-                 std::string_view /* log_level_description */,
-                 std::string_view /* log_level_short_code */,
+  void write_log(MacroMetadata const* /* log_metadata */, uint64_t /* log_timestamp */,
+                 std::string_view /* thread_id */, std::string_view /* thread_name */,
+                 std::string const& /* process_id */, std::string_view /* logger_name */, LogLevel log_level,
+                 std::string_view /* log_level_description */, std::string_view /* log_level_short_code */,
                  std::vector<std::pair<std::string, std::string>> const* /* named_args */,
                  std::string_view log_message, std::string_view log_statement) override
   {